moved class into own file
